meson options: vcs-tag=false
PKGBUILD clones the systemd repository instead of fetching a release
tarball. When built in a repository, git-describe output, complete with
a '^' dirty mark, is appended to the systemd version string by default.

The '^' dirty mark, in particular, can be problematic for tools that
(attempt to) parse the systemd version string. I ran into this problem
running `mkosi`, which does not function with a dirty systemd version
string formatted as such.

Before:
    /usr/lib/systemd/systemd --version | head -n 1
    systemd 256 (256.1-1-arch-g34ba18b^)

After:
    LD_LIBRARY_PATH=$PWD/pkg/systemd/usr/lib/systemd ./pkg/systemd/usr/lib/systemd/systemd --version | head -n 1
    systemd 256 (256.1-2-arch)

Signed-off-by: Joe Konno <joe.konno@intel.com>