Use vmlinux.h from linux-headers
Instead of generating vmlinux.h from whatever kernel happens to be
running on the build machine, let's pick it up from the linux-headers
package. This makes builds more reproducible and less dependent on the
build machine. It also makes it possible to build system in build
environments without /sys available.